worker/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "worker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2",
    "langchain-mcp-adapters>=0.0.5",
    "langchain-anthropic>=0.3",
    "langchain-core>=0.3",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "redis>=5.0",
    "pydantic-settings>=2.0",
    "langgraph-checkpoint-postgres>=2.0",
    "python-json-logger>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "simple-crewai-pair-agent @ file:../agents/simple_crewai_pair_agent",
    "httpx[asyncio]>=0.27",
    "gitpython>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "fakeredis[aioredis]>=2.0",
    "ruff>=0.4",
    "aiosqlite>=0.20",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: pure unit tests, no I/O",
    "integration: requires compiled graph",
]

[tool.ruff]
src = ["src"]
line-length = 100
